DB-1174: delete RPKI objects that are not referenced from manifests
RPKI objects that are (no longer) referenced from current manifests
are deleted after at least 7 days of unreachability.
