+++ src/main/resources/application.properties
@@ -59,4 +59,7 @@ rpki.validator.rsync.repository.download.interval=PT10M
 
 rpki.validator.rrdp.trust.all.tls.certificates=true
 
+rpki.validator.rpki.object.cleanup.interval.ms=3600000
+rpki.validator.rpki.object.cleanup.grace.duration=P7D
+
 spring.jackson.date-format=yyyy-MM-dd hh:mm:ss
